Simplify/Improve record serial management
Right now, the serial is bumped N times for every API call, which is
really not great when you consider that systemd-networkd and others like
it are caching name resolvers.

This can be elided by checking if a matching record exists in the
catalog before inserting it.

Currently this patch has / may have the following problems:

- [ ] May have a windowing issue for existing records that change over
      time. May just need a TTL refresh.
- [ ] PTR records still use the old method

Signed-off-by: Erik Hollensbe <linux@hollensbe.org>