WIP: Update/clean up dependencies, move API crates into repo, replace Docker w/Nix
This does a number of things which _could_ be separated, so I consider this more or
less experimental.

Implied in the removal of the various dockerfiles is the expectation that production
deployments of ZeroNSD will simply use a Docker container, rather than distro-native
packages.

TODO: automate publishing of containers from Nix-generated image